requirements.txt
rich
pyfiglet
paramiko
fpdf
requests
mysql-connector-python
dnspython
python-whois
